requirements.txt
aerich==0.5.8
aiomysql==0.0.21
aiosqlite==0.17.0
anyio==3.3.3
asgiref==3.4.1
asyncpg==0.24.0
atlastk==0.12.10
bcrypt==3.2.0
cffi==1.14.6
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
ddlparse==1.10.0
dictdiffer==0.9.0
dnspython==2.1.0
ecdsa==0.14.1
email-validator==1.1.3
fastapi==0.70.0
fastapi-pagination==0.9.0
h11==0.12.0
idna==3.2
iso8601==0.1.16
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
PyMySQL==0.9.3
pyparsing==2.4.7
pypika-tortoise==0.1.1
python-dotenv==0.19.1
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
tortoise-orm==0.17.8
typing-extensions==3.10.0.2
uvicorn==0.15.0
